🧪 WCARQG routing: isolate auto route E2E coverage
