+++ .agentplane/tasks/202608061646-WCARQG/pr/review.md
@@ -30,10 +30,10 @@ Created: 2026-08-06T16:48:32.143Z
 
 ```text
  docs/user/cli-reference.generated.mdx              |   1 +
- .../src/backends/task-backend/shared/record.ts     |  30 ++++++
+ .../src/backends/task-backend/shared/record.ts     |  30 +++++
  .../src/backends/task-backend/shared/types.ts      |   2 +
- .../src/cli/run-cli.core.route-decision.test.ts    | 104 +++++++++++++++++++
- .../src/cli/run-cli.core.tasks.create.test.ts      |  44 ++++++++
+ .../src/cli/run-cli.core.task-routing.test.ts      | 134 +++++++++++++++++++++
+ .../src/cli/run-cli.core.tasks.create.test.ts      |  44 +++++++
  .../src/commands/blueprint/task-input.test.ts      |  21 ++++
  .../src/commands/blueprint/task-input.ts           |   3 +-
  .../src/commands/branch/work-start.command.ts      |   8 +-
@@ -52,21 +52,21 @@ Created: 2026-08-06T16:48:32.143Z
  .../src/runtime/task-intake/resolve-normalize.ts   |   1 +
  .../agentplane/src/runtime/task-intake/types.ts    |   2 +
  .../agentplane/src/runtime/task-routing/index.ts   |   5 +
- .../src/runtime/task-routing/resolve.test.ts       |  75 ++++++++++++++
- .../agentplane/src/runtime/task-routing/resolve.ts | 115 +++++++++++++++++++++
- .../schemas/task-readme-frontmatter.schema.json    |  42 ++++++++
- packages/core/schemas/tasks-export.schema.json     |  42 ++++++++
+ .../src/runtime/task-routing/resolve.test.ts       |  75 ++++++++++++
+ .../agentplane/src/runtime/task-routing/resolve.ts | 115 ++++++++++++++++++
+ .../schemas/task-readme-frontmatter.schema.json    |  42 +++++++
+ packages/core/schemas/tasks-export.schema.json     |  42 +++++++
  packages/core/src/tasks/index.ts                   |   3 +
- .../core/src/tasks/task-artifact-schema.task.ts    |  12 +++
+ .../core/src/tasks/task-artifact-schema.task.ts    |  12 ++
  .../src/tasks/task-provider-safe-projection.ts     |   3 +
  packages/core/src/tasks/task-readme.ts             |   1 +
- packages/core/src/tasks/task-store.ts              |  12 +++
+ packages/core/src/tasks/task-store.ts              |  12 ++
  packages/core/src/tasks/tasks-export.ts            |   3 +
- .../schemas/task-readme-frontmatter.schema.json    |  42 ++++++++
- packages/spec/schemas/tasks-export.schema.json     |  42 ++++++++
- schemas/task-readme-frontmatter.schema.json        |  42 ++++++++
- schemas/tasks-export.schema.json                   |  42 ++++++++
- 37 files changed, 784 insertions(+), 9 deletions(-)
+ .../schemas/task-readme-frontmatter.schema.json    |  42 +++++++
+ packages/spec/schemas/tasks-export.schema.json     |  42 +++++++
+ schemas/task-readme-frontmatter.schema.json        |  42 +++++++
+ schemas/tasks-export.schema.json                   |  42 +++++++
+ 37 files changed, 814 insertions(+), 9 deletions(-)
 ```
 
 </details>
